Fine-tune message fetching. Reorganize flow types
